Refactor GraphQL API; move schema to single file
